spring data, all classes

+++ hypermarket-core/src/main/java/ro/sda/hypermarket/core/service/DepartmentServiceImpl.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ro.sda.hypermarket.core.dao.DepartmentDAO;
 import ro.sda.hypermarket.core.entity.Department;
+import ro.sda.hypermarket.core.repository.DepartmentRepository;
 
 @Service
 @Transactional(readOnly = true, rollbackFor = Exception.class)
@@ -13,25 +14,41 @@ public class DepartmentServiceImpl implements DepartmentService{
     @Autowired
     private DepartmentDAO departmentDAO;
 
+    @Autowired
+    private DepartmentRepository departmentRepository;
+
 
     @Override
-    public void createDepartment(Department department) {
-        departmentDAO.createDepartment(department);
+    public Department createDepartment(Department department, boolean useHibernate) {
+        if(useHibernate){
+            return  departmentDAO.createDepartment(department);
+        }
+            return departmentRepository.save(department);
     }
 
     @Override
-    public Department readDepartment(Long departmentId) {
-        return departmentDAO.readDepartment(departmentId);
+    public Department readDepartment(Long departmentId, boolean useHibernate) {
+        if(useHibernate){
+            return departmentDAO.readDepartment(departmentId);
+        }
+        return departmentRepository.findById(departmentId);
     }
 
     @Override
-    public void updateDepartment(Department department) {
-        departmentDAO.updateDepartment(department);
+    public Department updateDepartment(Department department, boolean useHibernate) {
+        if(useHibernate){
+            return departmentDAO.updateDepartment(department);
+        }
+        return departmentRepository.save(department);
     }
 
     @Override
-    public Department deleteDepartment(Long departmentId) {
-        return departmentDAO.deleteDepartment(departmentId);
+    public Department deleteDepartment(Long departmentId, boolean useHibernate) {
+        if(useHibernate){
+            departmentDAO.deleteDepartment(departmentId);
+        }
+            departmentRepository.delete(departmentId);
+        return null;
     }
