+++ hypermarket-core/src/main/java/ro/sda/hypermarket/core/service/SaleProductServiceImpl.java
@@ -5,6 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ro.sda.hypermarket.core.dao.SaleProductDAO;
 import ro.sda.hypermarket.core.entity.SaleProduct;
+import ro.sda.hypermarket.core.repository.SaleProductRepository;
 
 @Service
 @Transactional(readOnly = true, rollbackFor = Exception.class)
@@ -12,24 +13,38 @@ public class SaleProductServiceImpl implements SaleProductService {
 
     @Autowired
     private SaleProductDAO saleProductDAO;
+    @Autowired
+    private SaleProductRepository saleProductRepository;
 
     @Override
-    public void createSaleProduct(SaleProduct saleProduct) {
-        saleProductDAO.createSaleProduct(saleProduct);
+    public void createSaleProduct(SaleProduct saleProduct, boolean useHibernate) {
+        if(useHibernate){
+            saleProductDAO.createSaleProduct(saleProduct);
+        }
+            saleProductRepository.save(saleProduct);
     }
 
     @Override
-    public SaleProduct readSaleProduct(Long saleProductId) {
-        return saleProductDAO.readSaleProduct(saleProductId);
+    public SaleProduct readSaleProduct(Long saleProductId, boolean useHibernate) {
+        if(useHibernate){
+            return saleProductDAO.readSaleProduct(saleProductId);
+        }
+            return saleProductRepository.findById(saleProductId);
     }
 
     @Override
-    public SaleProduct updateSaleProduct(SaleProduct saleProduct) {
-        return saleProductDAO.updateSaleProduct(saleProduct);
+    public SaleProduct updateSaleProduct(SaleProduct saleProduct, boolean useHibernate) {
+        if(useHibernate){
+            return saleProductDAO.updateSaleProduct(saleProduct);
+        }
+            return saleProductRepository.save(saleProduct);
     }
 
     @Override
-    public void deleteSaleProduct(SaleProduct saleProduct) {
-        saleProductDAO.deleteSaleProduct(saleProduct);
+    public void deleteSaleProduct(SaleProduct saleProduct, boolean useHibernate) {
+        if(useHibernate){
+            saleProductDAO.deleteSaleProduct(saleProduct);
+        }
+            saleProductRepository.delete(saleProduct);
     }
 }
